+++ src/metrepl/middleware/tests_metrics.clj
@@ -5,17 +5,33 @@
    [metrepl.transport :as m.transport]
    [nrepl.middleware :as middleware]))
 
+(defn ^:private test-results-summary [results]
+  (into {}
+        (for [[ns m] results
+              [var v] m]
+          [(keyword (name ns) (name var))
+           v])))
+
 (defn ^:private handle-test [response]
   (when-let [summary (some-> (get response "summary") walk/keywordize-keys)]
-    (let [results (get response "results")]
-      (metrics/metrify :event/test-executed
+    (let [results (walk/keywordize-keys (get response "results"))
+          tests (test-results-summary results)]
+      (metrics/metrify :event/tests-executed
                        {:elapsed-time-ms (get-in response ["elapsed-time" "ms"])
                         :ns (keys results)
-                        :summary (walk/keywordize-keys summary)}))))
+                        :summary (walk/keywordize-keys summary)})
+      (doseq [[_var tests] tests]
+        (doseq [test tests]
+          (when-let [event-name (case (:type test)
+                                  "pass" :event/test-passed
+                                  "fail" :event/test-failed
+                                  "error" :event/test-errored
+                                  nil)]
+            (metrics/metrify event-name (select-keys test [:ns :var :file :error :context]))))))))
 
 (defn wrap-tests-metrics
   "Wrap test related ops metrifying, emmiting these events:
-   - `:event/test-executed` when test finished successfully or failed."
+   - `:event/tests-executed` when test finished successfully or failed."
   [handler]
   (fn [{:keys [op] :as msg}]
     (if (= "test" op)